+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import { supabase } from '../utils/supabaseClient';
 import Auth from '../components/Auth';
-import Account from '../components/Account';
+import Account from '../components/organisms/users/Account';
 import { Session } from '@supabase/supabase-js';
 import { Box, Flex } from '@chakra-ui/react';
 export default function Home() {
